requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
torch
transformers
pdf2image
pdfplumber
sentence-transformers
faiss-cpu
rank-bm25
numpy
nltk